+++ ruby/demo_1.rb
@@ -3,18 +3,18 @@
 require 'yaml'
 require 'byebug'
 
-require './mf_client'
+require './mediaflux'
 
-# Hash with symbolic keys; Avoid extra Rails dependencies for now: 
+# Load hash with symbolic keys; Avoid extra Rails dependencies for now: 
 config = YAML.load_file('config.yaml').map { |k, v| [k.to_sym, v] }.to_h
 
-mf = MFClient.new(**config)
+mf = MediaFlux::MFClient.new(**config)
 
 
 begin
   puts "\nUnknown service:"
   mf.call("foo.bar")
-rescue MFError => e
+rescue MediaFlux::MFError => e
   puts e.error
   puts e.message
   puts e.stack.split("\n").first
@@ -65,27 +65,7 @@ mf.session() do
     # puts "\nDelete:"
     # doc = mf.call("asset.destroy", id: id)
     # puts doc
-  rescue MFError => e
-    puts e.message
-  end
-
-  begin
-    puts "\nImport asset by reference FAILS:"
-    doc = mf.call("asset.import", url: {
-      _by: "reference",
-      _: "https://www.princeton.edu/robots.txt"
-    })
-    # puts doc
-    # id = doc.elements["//id"].first
-
-    # puts "\nGet asset:"
-    # doc = mf.call("asset.get", id: id)
-    # puts doc
-
-    # puts "\nDelete:"
-    # doc = mf.call("asset.destroy", id: id)
-    # puts doc
-  rescue MFError => e
+  rescue MediaFlux::MFError => e
     puts e.message
   end
 end
